telegram_bot/requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.0
httpx==0.27.0